Update fr.po
I've updated a few elements of this translation as requested. I also noticed that the old translation (before my modifications in this new request to update the French translation file) was not appearing correctly in the application. I'm running Fedora 40 and it concerns paragraph #: schemas/org.gnome.shell.extensions.soft-brightness-plus.gschema.xml:53
msgid “” which is currently displayed in English even in the French application.